Split up old chapter The extension file.
There should be no more very large code fences
in existance. Much of the code has been divided in
a group of sections, chapters. The represent some
order, but I do believe this text can do with a
couple of more revisions so that it becomes even
more readable and explains every bit and piece in
enough detail, giving all necessary argumentation
so that one can always understand what is happening
and for what reason.