Introduce ending token for starred fragment filename.
Fix #34 through introduction of ending token, which is essentially
whitespace and a dollar sign.

The reason to do this is to also prepare for extra information on
the options line after the filename. For instance it now becomes
easy to add information about say templates to use for a file.