Add simple mermaid.js support.
Code fences with mermaid diagrams expressed
are rendered in the renderCodeFence in such
a way that they can be handled by the mermaid
module imported in the default HTML template.

This means also that custom templates need to
import mermaid to be able to render diagrams
should the author wish to do so.

Fixes #9